requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.4.1=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py312hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.7.4=h56e8100_0
certifi=2024.7.4=pyhd8ed1ab_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
contourpy=1.2.1=pypi_0
cuda-cccl=12.5.39=0
cuda-cccl_win-64=12.5.39=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.5.39=0
cuda-opencl-dev=12.5.39=0
cuda-profiler-api=12.5.39=0
cuda-runtime=12.1.0=0
cuda-version=12.5=3
cycler=0.12.1=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.8=pypi_0
et-xmlfile=1.1.0=pypi_0
executing=2.0.1=pypi_0
expat=2.6.2=hd77b12b_0
filelock=3.13.1=py312haa95532_0
fonttools=4.53.0=pypi_0
freetype=2.12.1=ha860e81_0
idna=3.7=py312haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.4=pypi_0
ipython=8.25.0=pypi_0
ipywidgets=8.1.3=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=py312haa95532_0
joblib=1.4.2=pypi_0
jpeg=9e=h2bbff1b_1
jupyter-client=8.6.2=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-widgets=3.0.11=pypi_0
kiwisolver=1.4.5=pypi_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libblas=3.9.0=1_h8933c1f_netlib
libcblas=3.9.0=5_hd5c7e75_netlib
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.6.39=0
libcurand-dev=10.3.6.39=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libexpat=2.6.2=h63175ca_0
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
liblapack=3.9.0=5_hd5c7e75_netlib
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.39=h8cc25b3_0
libsqlite=3.46.0=h2466b09_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp-base=1.3.2=h2bbff1b_0
libzlib=1.2.13=h2466b09_6
lz4-c=1.9.4=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.3=py312h2bbff1b_0
matplotlib=3.9.0=pypi_0
matplotlib-inline=0.1.7=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py312h2bbff1b_1
mkl_fft=1.3.8=py312h2bbff1b_0
mkl_random=1.2.4=py312h59b6b97_0
mpmath=1.3.0=py312haa95532_0
msys2-conda-epoch=20160418=1
nest-asyncio=1.6.0=pypi_0
networkx=3.1=py312haa95532_0
numpy=1.26.4=py312hfd52020_0
numpy-base=1.26.4=py312h4dde369_0
opencv-python=4.10.0.84=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openpyxl=3.1.5=pypi_0
openssl=3.3.1=h2466b09_2
packaging=24.0=pypi_0
pandas=2.2.2=py312h72972c8_1
parso=0.8.4=pypi_0
patsy=0.5.6=pyhd8ed1ab_0
pillow=10.3.0=py312h2bbff1b_0
pip=24.0=py312haa95532_0
platformdirs=4.2.2=pypi_0
prompt-toolkit=3.0.45=pypi_0
psutil=5.9.8=pypi_0
pure-eval=0.2.2=pypi_0
pydmd=1.0.0=pypi_0
pygments=2.18.0=pypi_0
pypalettes=0.1.2=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py312haa95532_0
python=3.12.3=h2628c8c_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.12=4_cp312
pytorch=2.3.0=py3.12_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pyhd8ed1ab_0
pywin32=306=pypi_0
pyyaml=6.0.1=py312h2bbff1b_0
pyzmq=26.0.3=pypi_0
requests=2.32.2=py312haa95532_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=69.5.1=py312haa95532_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.45.3=h2bbff1b_0
stack-data=0.6.3=pypi_0
statsmodels=0.14.2=py312h1a27103_0
sympy=1.12=py312haa95532_0
tbb=2021.8.0=h59b6b97_0
tdqm=0.0.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h0416ee5_0
torchaudio=2.3.0=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.18.0=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pypi_0
typing_extensions=4.11.0=py312haa95532_0
tzdata=2024a=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.2.1=py312haa95532_0
vc=14.2=h2eaa2aa_1
vc14_runtime=14.40.33810=ha82c5b3_20
vs2015_runtime=14.40.33810=h3bf8584_20
wcwidth=0.2.13=pypi_0
wheel=0.43.0=py312haa95532_0
widgetsnbextension=4.0.11=pypi_0
win_inet_pton=1.1.0=py312haa95532_0
xlsxwriter=3.2.0=pypi_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
zlib=1.2.13=h2466b09_6
zstd=1.5.5=hd43e919_2
